+++ src/main/java/com/digital/Clase23ClinicaOdontologica/repository/IDao.java
@@ -6,8 +6,8 @@ public interface IDao<T>{
     //alta, buscarID, eliminar,modificar
 
     T guardar(T t);
-    T buscarID(int id);
-    void eliminar (int id);
+    T buscarID(Long id);
+    void eliminar (Long id);
     void modificar(T t);
     List<T> listar();
     T buscarXString(String valor);
